# contracts/openapi/finance-api.v1.yaml
openapi: 3.1.0
...
info:
  title: PFOS finance-api
  version: 1.0.0-draft.1
  summary: Personal Finance Operating System — REST API v1 (draft)
  description: |
    REST API of the Personal Finance Operating System (PFOS).

    * All business resources live under `/workspaces/{workspaceId}`.
    * Money is always `{ "amount": "<decimal string>", "currency": "<code>" }`. Never JSON numbers.
    * Business dates are `YYYY-MM-DD` (workspace timezone); instants are RFC 3339 UTC.
    * Errors are RFC 9457 `application/problem+json` with a stable domain `code`.
    * Financial `POST` operations require `Idempotency-Key`; updates and actions on existing aggregates require `If-Match`.
    * Every operation declares `x-openspec-capability` (OpenSpec taxonomy, ARCHITECTURE §14) and `x-required-role`.

    Only Phase 1 paths are specified. Later-phase resources are listed as tags (description only).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  contact:
    name: PFOS owner
servers:
  - url: /api/v1
    description: Relative base path (served behind the finance-web BFF)
security:
  - bearerAuth: []
...
paths:
...
  /me:
    get:
      tags: [Identity]
      operationId: getMe
      summary: Get the authenticated user and their workspace memberships
      x-openspec-capability: identity/authentication
      x-required-role: AUTHENTICATED
      responses:
        '200':
          description: Current user
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    patch:
      tags: [Identity]
      operationId: updateMe
      summary: Update the authenticated user's profile and preferences
      x-openspec-capability: identity/authentication
      x-required-role: AUTHENTICATED
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/MeUpdate' }
      responses:
        '200':
          description: Updated user
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
  /workspaces/{workspaceId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Identity]
      operationId: getWorkspace
      summary: Get a workspace
      x-openspec-capability: identity/workspace-membership
      x-required-role: VIEWER
      responses:
        '200':
          description: Workspace
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      tags: [Identity]
      operationId: updateWorkspace
      summary: Update workspace settings (name, timezone, locale, base currency)
      x-openspec-capability: identity/workspace-membership
      x-required-role: OWNER
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/WorkspaceUpdate' }
      responses:
        '200':
          description: Updated workspace
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }

  # ───────────────────────────── Accounts ─────────────────────────────
  /workspaces/{workspaceId}/accounts:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Accounts]
      operationId: listAccounts
      summary: List accounts with current balances
      x-openspec-capability: accounts/account-management
      x-required-role: VIEWER
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/IncludeArchived'
        - name: type
          in: query
          description: Filter by account type (repeatable)
          schema:
            type: array
            items: { $ref: '#/components/schemas/AccountType' }
          style: form
          explode: true
        - name: currency
          in: query
          description: Filter by currency code (repeatable)
          schema:
            type: array
            items: { $ref: '#/components/schemas/CurrencyCode' }
          style: form
          explode: true
        - name: institutionId
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: sort
          in: query
          description: Sort keys. Default `displayOrder`.
          schema:
            type: string
            enum: [displayOrder, name, -name, createdAt, -createdAt]
            default: displayOrder
      responses:
        '200':
          description: Accounts page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [Accounts]
      operationId: createAccount
      summary: Create an account (optionally with an opening balance, which posts an OPENING_BALANCE transaction)
      x-openspec-capability: [accounts/account-management, ledger/journal-posting]
      x-required-role: EDITOR
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '201':
          description: Account created
          headers:
            Location: { $ref: '#/components/headers/Location' }
            ETag: { $ref: '#/components/headers/ETag' }
            Idempotent-Replayed: { $ref: '#/components/headers/IdempotentReplayed' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /workspaces/{workspaceId}/transactions:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: List transactions (legs and splits inline)
      x-openspec-capability: transactions/transaction-recording
      x-required-role: VIEWER
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Query'
        - name: accountId
          in: query
          description: Transactions with at least one leg on any of these accounts
          schema:
            type: array
            items: { $ref: '#/components/schemas/Uuid' }
          style: form
          explode: true
        - name: categoryId
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/Uuid' }
          style: form
          explode: true
        - name: tagId
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/Uuid' }
          style: form
          explode: true
        - name: counterpartyId
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/Uuid' }
          style: form
          explode: true
        - name: kind
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/TransactionKind' }
          style: form
          explode: true
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/TransactionStatus' }
          style: form
          explode: true
        - name: currency
          in: query
          schema: { $ref: '#/components/schemas/CurrencyCode' }
        - name: dateFrom
          in: query
          description: Inclusive lower bound on transactionDate
          schema: { $ref: '#/components/schemas/LocalDate' }
        - name: dateTo
          in: query
          description: Inclusive upper bound on transactionDate
          schema: { $ref: '#/components/schemas/LocalDate' }
        - name: amountMin
          in: query
          description: Inclusive lower bound on the transaction amount magnitude
          schema: { $ref: '#/components/schemas/PositiveDecimalString' }
        - name: amountMax
          in: query
          description: Inclusive upper bound on the transaction amount magnitude
          schema: { $ref: '#/components/schemas/PositiveDecimalString' }
        - name: sort
          in: query
          description: Default `-transactionDate` (ties broken by id).
          schema:
            type: string
            enum: [transactionDate, -transactionDate, amount, -amount, createdAt, -createdAt]
            default: -transactionDate
      responses:
        '200':
          description: Transactions page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Record an income, expense, refund or adjustment with splits
      description: |
        Transfers and conversions have dedicated endpoints. Amounts in commands are positive magnitudes;
        direction is given by `kind`. If `splits` is omitted a single split in the system category
        UNCATEGORIZED is created. Non-PENDING transactions are posted to the ledger in the same DB transaction.
      x-openspec-capability: [transactions/transaction-recording, transactions/splits, ledger/journal-posting]
      x-required-role: EDITOR
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionCreate' }
            examples:
              splitExpense:
                summary: Supermarket expense split in two categories
                value:
                  kind: EXPENSE
                  transactionDate: '2026-09-30'
                  accountId: 0192f3c4-7b2e-7c1a-9d8e-3f2a1b0c9d01
                  amount: { amount: '350.00', currency: BOB }
                  description: Hipermaxi
                  splits:
                    - amount: { amount: '300.00', currency: BOB }
                      categoryId: 0192f3c4-7b2e-7c1a-9d8e-3f2a1b0c9d11
                    - amount: { amount: '50.00', currency: BOB }
                      categoryId: 0192f3c4-7b2e-7c1a-9d8e-3f2a1b0c9d12
                      tagIds: [0192f3c4-7b2e-7c1a-9d8e-3f2a1b0c9d21]
      responses:
        '201':
          description: Transaction recorded
          headers:
            Location: { $ref: '#/components/headers/Location' }
            ETag: { $ref: '#/components/headers/ETag' }
            Idempotent-Replayed: { $ref: '#/components/headers/IdempotentReplayed' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC access token (JWT) issued by the workspace IdP (Keycloak locally). Obtained by the finance-web BFF via
        Authorization Code + PKCE and forwarded server-side; never exposed to browser JavaScript.
        The API validates signature (JWKS), `iss`, `aud`, `exp`, `nbf`.
        IdP discovery: `<issuer>/.well-known/openid-configuration` (issuer configured per environment;
        locally Keycloak realm `pfos`). An `openIdConnect` scheme is not declared because the browser never
        talks to the API directly (BFF pattern, ADR-0010/ADR-0019).
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Workspace UUIDv7. Caller must be an active member.
      schema: { $ref: '#/components/schemas/Uuid' }
...
    Limit:
      name: limit
      in: query
      description: Page size
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from `page.nextCursor`. Valid only for the same filters and sort.
      schema: { type: string, maxLength: 1024 }
    Query:
      name: q
      in: query
      description: Free-text search
      schema: { type: string, minLength: 1, maxLength: 200 }
    IncludeArchived:
      name: includeArchived
      in: query
      schema: { type: boolean, default: false }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Required on financial POSTs. Same key + same payload within retention replays the stored response
        (`Idempotent-Replayed: true`); same key + different payload → 422 IDEMPOTENCY_KEY_REUSED;
        key still in progress → 409 IDEMPOTENCY_REQUEST_IN_PROGRESS.
      schema: { $ref: '#/components/schemas/IdempotencyKeyValue' }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag of the version being modified. Missing → 428; mismatch → 412.
      schema: { type: string, examples: ['"7"'] }
...
  headers:
    ETag:
      description: Strong entity tag derived from the aggregate `version`
      schema: { type: string, examples: ['"3"'] }
    Location:
      description: URI of the created resource or of the operation to poll
      schema: { type: string, format: uri-reference }
    IdempotentReplayed:
      description: Present and `true` when the response is a replay of a stored idempotent response
      schema: { type: boolean }
    RetryAfter:
      description: Seconds to wait before retrying / polling again
      schema: { type: integer, minimum: 0 }
    RateLimit:
      description: 'Remaining quota (IETF RateLimit header draft), e.g. `"default";r=412;t=23`'
      schema: { type: string }
    RateLimitPolicy:
      description: 'Quota policy (IETF RateLimit header draft), e.g. `"default";q=600;w=60`'
      schema: { type: string }
  responses:
...
    BadRequest:
      description: Malformed request or schema validation failure (VALIDATION_FAILED, INVALID_CURSOR, INVALID_SORT, INVALID_FILTER)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid access token (UNAUTHENTICATED)
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Not an active member of the workspace (WORKSPACE_ACCESS_DENIED) or role too low (INSUFFICIENT_ROLE)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: Domain state conflict (e.g. PERIOD_CLOSED, ACCOUNT_ARCHIVED, INVALID_STATUS_TRANSITION, CONCURRENCY_CONFLICT, NAME_TAKEN, IDEMPOTENCY_REQUEST_IN_PROGRESS)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionFailed:
      description: If-Match does not match the current version (PRECONDITION_FAILED)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionRequired:
      description: Missing If-Match (PRECONDITION_REQUIRED) or Idempotency-Key (IDEMPOTENCY_KEY_REQUIRED)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnprocessableEntity:
      description: Domain validation failed (e.g. SPLITS_DO_NOT_SUM, AMOUNT_SCALE_EXCEEDED, CURRENCY_MISMATCH, REFERENCE_NOT_FOUND, IDEMPOTENCY_KEY_REUSED)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Rate limit exceeded (RATE_LIMITED)
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        RateLimit: { $ref: '#/components/headers/RateLimit' }
        RateLimit-Policy: { $ref: '#/components/headers/RateLimitPolicy' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Uuid:
      type: string
      format: uuid
      description: UUIDv7
      examples: [0192f3c4-7b2e-7c1a-9d8e-3f2a1b0c9d8e]
    IdempotencyKeyValue:
      type: string
      minLength: 16
      maxLength: 128
      pattern: '^[A-Za-z0-9_-]+$'
...
    PositiveDecimalString:
      type: string
      description: Strictly positive decimal string (validated > 0 by the server)
      pattern: '^(0|[1-9][0-9]{0,19})(\.[0-9]{1,18})?$'
      examples: ['685.00']
    CurrencyCode:
      type: string
      pattern: '^[A-Z0-9][A-Z0-9_.-]{1,15}$'
      examples: [BOB, USD, USDT, BTC]
    LocalDate:
      type: string
      format: date
      description: Business date (YYYY-MM-DD) in the workspace timezone
      examples: ['2026-09-30']
    Instant:
      type: string
      format: date-time
      description: RFC 3339 UTC instant
      examples: ['2026-09-30T18:42:00Z']
    Version:
      type: integer
      minimum: 1
      description: Optimistic locking version of the aggregate (also exposed as ETag)
    Money:
      type: object
      description: Monetary amount. `amount` uses the canonical scale of `currency`.
      additionalProperties: false
      required: [amount, currency]
      properties:
        amount: { $ref: '#/components/schemas/DecimalString' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
    PositiveMoney:
      type: object
      description: Monetary magnitude used in commands (must be > 0)
      additionalProperties: false
      required: [amount, currency]
      properties:
        amount: { $ref: '#/components/schemas/PositiveDecimalString' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
...
    PageInfo:
      type: object
      required: [limit, hasMore]
      properties:
        limit: { type: integer }
        hasMore: { type: boolean }
        nextCursor: { type: [string, 'null'] }
...
    Problem:
      type: object
      description: RFC 9457 problem details with PFOS extensions
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          examples: ['https://pfos.dev/problems/splits-do-not-sum']
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { $ref: '#/components/schemas/ErrorCode' }
        requestId: { type: string }
        currentVersion:
          type: integer
          description: Present on PRECONDITION_FAILED / CONCURRENCY_CONFLICT
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
...
    WorkspaceRole:
      type: string
      enum: [OWNER, EDITOR, VIEWER]
    Me:
      type: object
      required: [id, email, displayName, locale, memberships, version]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        email: { type: string, format: email }
        displayName: { type: string }
        locale: { type: string, examples: [es-BO] }
        preferences:
          type: object
          additionalProperties: true
          description: Free-form UI preferences (no financial data)
        memberships:
          type: array
          items:
            type: object
            required: [workspaceId, workspaceName, role]
            properties:
              workspaceId: { $ref: '#/components/schemas/Uuid' }
              workspaceName: { type: string }
              role: { $ref: '#/components/schemas/WorkspaceRole' }
        version: { $ref: '#/components/schemas/Version' }
    MeUpdate:
      type: object
      additionalProperties: false
      properties:
        displayName: { type: string, minLength: 1, maxLength: 100 }
        locale: { type: string, maxLength: 20 }
        preferences: { type: object, additionalProperties: true }
    Workspace:
      type: object
      required: [id, name, baseCurrency, timezone, locale, status, role, version, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        baseCurrency: { $ref: '#/components/schemas/CurrencyCode' }
        timezone: { type: string, examples: [America/La_Paz] }
        locale: { type: string, examples: [es-BO] }
        status:
          type: string
          enum: [ACTIVE, PENDING_DELETION]
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        version: { $ref: '#/components/schemas/Version' }
        createdAt: { $ref: '#/components/schemas/Instant' }
        updatedAt: { $ref: '#/components/schemas/Instant' }
...
    WorkspaceUpdate:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        timezone: { type: string }
        locale: { type: string }
        baseCurrency:
          $ref: '#/components/schemas/CurrencyCode'
...
    AccountType:
      type: string
      enum: [CASH, CHECKING, SAVINGS, CREDIT_CARD, LOAN, CRYPTO_WALLET, EXCHANGE, P2P, INVESTMENT, OTHER_ASSET, OTHER_LIABILITY]
    AccountClassification:
      type: string
      enum: [ASSET, LIABILITY]
    Account:
      type: object
      required: [id, name, type, classification, currency, balance, includeInNetWorth, version, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        type: { $ref: '#/components/schemas/AccountType' }
        classification: { $ref: '#/components/schemas/AccountClassification' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        institutionId: { type: [string, 'null'], format: uuid }
        openedOn: { type: [string, 'null'], format: date }
        closedOn: { type: [string, 'null'], format: date }
        balance:
          $ref: '#/components/schemas/Money'
          description: Presented balance (ASSET as is; LIABILITY as positive amount owed). Includes posted, cleared and reconciled.
        clearedBalance:
          $ref: '#/components/schemas/Money'
          description: Balance of cleared + reconciled transactions only
        pendingAmount:
          $ref: '#/components/schemas/Money'
          description: Net amount of PENDING transactions (not in the ledger)
        includeInNetWorth: { type: boolean }
        includeInBudget: { type: boolean }
        displayOrder: { type: integer }
        accountNumberLast4: { type: [string, 'null'], pattern: '^[0-9]{4}$' }
        color: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        archivedAt: { type: [string, 'null'], format: date-time }
        version: { $ref: '#/components/schemas/Version' }
        createdAt: { $ref: '#/components/schemas/Instant' }
        updatedAt: { $ref: '#/components/schemas/Instant' }
    AccountCreate:
      type: object
      additionalProperties: false
      required: [name, type, currency]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string, minLength: 1, maxLength: 100 }
        type: { $ref: '#/components/schemas/AccountType' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        institutionId: { $ref: '#/components/schemas/Uuid' }
        openedOn: { $ref: '#/components/schemas/LocalDate' }
        openingBalance:
          type: object
          description: Posts an OPENING_BALANCE transaction against EQUITY:OPENING_BALANCE:<CCY>
          additionalProperties: false
          required: [amount, date]
          properties:
            amount:
              $ref: '#/components/schemas/Money'
              description: Presented balance (for LIABILITY, positive = amount owed). Must use the account currency.
            date: { $ref: '#/components/schemas/LocalDate' }
        includeInNetWorth: { type: boolean, default: true }
        includeInBudget: { type: boolean, default: true }
        accountNumberLast4: { type: string, pattern: '^[0-9]{4}$' }
        color: { type: string, maxLength: 20 }
        notes: { type: string, maxLength: 2000 }
...
    AccountPage:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Account' }
        page: { $ref: '#/components/schemas/PageInfo' }
...
    TransactionKind:
      type: string
      enum: [INCOME, EXPENSE, TRANSFER, REFUND, ADJUSTMENT, OPENING_BALANCE, CONVERSION, LOAN_DISBURSEMENT, LOAN_PAYMENT, CARD_PAYMENT]
    TransactionStatus:
      type: string
      enum: [PENDING, POSTED, CLEARED, RECONCILED, VOIDED]
    TransactionSource:
      type: string
      enum: [MANUAL, IMPORT, RECURRING, DEBT, GOAL, SYSTEM]
    TransactionLeg:
      type: object
      description: Money movement on a user account, signed from the account perspective (outflow negative).
      required: [accountId, amount, role]
      properties:
        accountId: { $ref: '#/components/schemas/Uuid' }
        amount: { $ref: '#/components/schemas/Money' }
        role:
          type: string
          enum: [MAIN, SOURCE, TARGET, FEE]
    TransactionSplit:
      type: object
      required: [id, amount, categoryId]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        amount:
          $ref: '#/components/schemas/Money'
          description: Positive magnitude; direction given by the transaction kind
        categoryId: { $ref: '#/components/schemas/Uuid' }
        counterpartyId: { type: [string, 'null'], format: uuid }
        tagIds:
          type: array
          items: { $ref: '#/components/schemas/Uuid' }
        memo: { type: [string, 'null'] }
        customFields:
          type: object
          description: Custom field values keyed by field key (Phase 2)
          additionalProperties:
            type: [string, number, boolean, 'null']
    SplitInput:
      type: object
      additionalProperties: false
      required: [amount, categoryId]
      properties:
        amount: { $ref: '#/components/schemas/PositiveMoney' }
        categoryId: { $ref: '#/components/schemas/Uuid' }
        counterpartyId: { $ref: '#/components/schemas/Uuid' }
        tagIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { $ref: '#/components/schemas/Uuid' }
        memo: { type: string, maxLength: 500 }
...
    ConversionDetail:
      type: object
      description: Immutable pricing metadata of a conversion (docs/09-ledger-design.md §7). Never recalculated with current rates.
      required: [sourceAccountId, targetAccountId, sourceAmount, convertedSourceAmount, grossTargetAmount, targetAmount, effectiveRate, fees, executedAt]
      properties:
        sourceAccountId: { $ref: '#/components/schemas/Uuid' }
        targetAccountId: { $ref: '#/components/schemas/Uuid' }
        sourceAmount:
          $ref: '#/components/schemas/Money'
          description: Gross amount delivered (includes fees deducted in source currency)
        convertedSourceAmount: { $ref: '#/components/schemas/Money' }
        grossTargetAmount: { $ref: '#/components/schemas/Money' }
        targetAmount:
          $ref: '#/components/schemas/Money'
          description: Net amount received
        quotedRate:
          oneOf:
            - $ref: '#/components/schemas/Rate'
            - type: 'null'
        effectiveRate: { $ref: '#/components/schemas/Rate' }
        referenceRate:
          oneOf:
            - $ref: '#/components/schemas/ReferenceRate'
            - type: 'null'
        spread:
          oneOf:
            - $ref: '#/components/schemas/Spread'
            - type: 'null'
        fees:
          type: array
          items: { $ref: '#/components/schemas/ConversionFee' }
        provider: { $ref: '#/components/schemas/ConversionProvider' }
        executedAt: { $ref: '#/components/schemas/Instant' }
        externalRef: { type: [string, 'null'] }
...
    Transaction:
      type: object
      required: [id, kind, status, transactionDate, amount, legs, splits, source, revision, version, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        kind: { $ref: '#/components/schemas/TransactionKind' }
        status: { $ref: '#/components/schemas/TransactionStatus' }
        transactionDate: { $ref: '#/components/schemas/LocalDate' }
        amount:
          $ref: '#/components/schemas/Money'
          description: Magnitude in the main/source account currency
        description: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        counterpartyId: { type: [string, 'null'], format: uuid }
        legs:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/TransactionLeg' }
        splits:
          type: array
          description: Current splits (nominal part). Empty for plain transfers/conversions without fees.
          items: { $ref: '#/components/schemas/TransactionSplit' }
        conversion:
          description: Present only for kind CONVERSION
          oneOf:
            - $ref: '#/components/schemas/ConversionDetail'
            - type: 'null'
        source: { $ref: '#/components/schemas/TransactionSource' }
        externalRef:
          oneOf:
            - $ref: '#/components/schemas/ExternalRef'
            - type: 'null'
        revision:
          type: integer
          minimum: 1
          description: Increments each time the transaction is re-posted to the ledger
        voidedAt: { type: [string, 'null'], format: date-time }
        voidReason: { type: [string, 'null'] }
        attachmentCount: { type: integer, minimum: 0 }
        version: { $ref: '#/components/schemas/Version' }
        createdAt: { $ref: '#/components/schemas/Instant' }
        updatedAt: { $ref: '#/components/schemas/Instant' }
    TransactionCreate:
      type: object
      additionalProperties: false
      required: [kind, transactionDate, accountId, amount]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        kind:
          type: string
          enum: [INCOME, EXPENSE, REFUND, ADJUSTMENT]
        direction:
          type: string
          enum: [INCREASE, DECREASE]
          description: Required only for ADJUSTMENT (effect on the presented account balance)
        status:
          type: string
          enum: [PENDING, POSTED, CLEARED]
          default: POSTED
        transactionDate: { $ref: '#/components/schemas/LocalDate' }
        accountId: { $ref: '#/components/schemas/Uuid' }
        amount:
          $ref: '#/components/schemas/PositiveMoney'
          description: Must use the account currency
        description: { type: string, maxLength: 500 }
        notes: { type: string, maxLength: 4000 }
        counterpartyId: { $ref: '#/components/schemas/Uuid' }
        splits:
          type: array
          minItems: 1
          maxItems: 50
          description: Must sum to `amount` (SPLITS_DO_NOT_SUM). Omit for a single UNCATEGORIZED split.
          items: { $ref: '#/components/schemas/SplitInput' }
...
    TransactionPage:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
        page: { $ref: '#/components/schemas/PageInfo' }

    # ── Classification ──
...